--- packages/doenetml-worker-rust/lib-doenetml-core/src/core/state_var_interfaces.rs
@@ -1,4 +1,0 @@
-pub mod boolean_state_var_interfaces;
-pub mod math_state_var_interfaces;
-pub mod text_state_var_interfaces;
-mod util;
